Expose init cert
